% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_fill.R
\name{xlsform_fill}
\alias{xlsform_fill}
\title{generate data for an xlsform}
\usage{
xlsform_fill(questions, choices, n)
}
\arguments{
\item{questions}{the xlsform questions sheet as data frame}

\item{choices}{the xlsform choices sheet as a data frame}

\item{n}{the number of rows to generate}
}
\description{
generate data for an xlsform
}
