% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_samplingframe.R
\name{xlsform_generate_samplingframe}
\alias{xlsform_generate_samplingframe}
\title{create sampling frame from kobo tool}
\usage{
xlsform_generate_samplingframe(choices, strata_list_names)
}
\arguments{
\item{choices}{the kobo questionnaire choices sheet as a data.frame}

\item{strata_list_names}{the choice list_name column values of the variables that together uniquely define each stratum}
}
\value{
a data frame with two columns:
"strata": the names of the strata; all unique combinations of the choices in each provided strata list name, concatenated by two underscores ("__")
"population": randomly generated population for each stratum (lognormal distribution with meanlog 10 and sdlog 1)
}
\description{
create sampling frame from kobo tool
}
